[Client] Improve stop() handling when it triggers PCSC shutdown errors (#2)
Ran into an edge case while adding stop/resume functionality to the E2E
app - if the client is is stopped while waiting on state updates, a
`NoService` error code is returned (at least on macos so far).

Was previously only ignoring the `Cancelled` code, so the app exits on
the unexpected error - switching to tracking the running state of the
client in JS and ignoring errors while we're explicitly shutting down.

Bumping patch version to v0.1.2.